black sidebar bug fixed, SdlWindow independet res
rendering made to always happen entirely on SdlWindow::refresh() call*,
as if an atomic operation, so wrong SDL_RenderPresent() calls can't be
made out of order and flush the renderer before everything has been
rendered. this was probably the black sidebar cause.
* the actual redraw only happens once per window event loop (this
prevents MapFrame::runNoise() from showing progress with the current
implementation), but fully redrawing at each refresh() may work although
testing is needed.
SdlWindow now has an independent resolution for the contents, with an
intermediate texture with this same resolution which contents are
rendered onto before scaling this one tex to the screen. this prevents
the discretization bug when resizing several textures that should be
seamless between each other.